GOF-23 Moved service functions, created assignment
Moved functions in firebaseService to seperate service files in a
service folder because the size was getting too large

Created service and database operations Create, Update, and Delete for
assignments as well as outline of api
