+++ src/app/api/organizations/[organizationId]/assignments/assignmentsPost.ts
@@ -1,0 +1,74 @@
+import { cookies } from "next/headers";
+import { NextRequest, NextResponse } from "next/server";
+
+import { FirebaseVerifyError } from "@/api/firebase/firebaseVerify";
+import { FirestoreDatabaseError } from "@/api/firebase/firestoreDatabase";
+import { assignmentSchema } from "@/api/database/database";
+import { addAssignmentToOrg } from "@/api/firebase/firebaseService/assignmentService";
+
+/**
+ * This route handles the creation of a assignment of a truck
+ * Path: POST /api/organizations/{organizationId}/assignments
+ */
+export async function assignmentsPOST(
+  request: NextRequest,
+  { organizationId }: { organizationId: string }
+) {
+  try {
+
+    // Validate the incoming request body against the assignment schema.
+    const requestBody = await request.json();
+    // Only need the truckId since rest of data comes from token
+    const assignmentTruckIdSchema = assignmentSchema.omit({ assignmentId: true, assignedAt: true });
+    const validationResult = assignmentTruckIdSchema.safeParse(requestBody);
+
+    // Check if validation failed
+    if (!validationResult.success) {
+      return NextResponse.json(
+        { status: "fail", message: validationResult.error.message },
+        { status: 400 } // Bad Request
+      );
+    }
+    const initialAssignmentData = validationResult.data;
+
+    // Get user token
+    const userCookies = await cookies();
+    const token = userCookies.get('session-token')?.value;
+    if (!token) {
+      return NextResponse.json(
+        { status: "fail", message: "Authentication required." },
+        { status: 401 } // Unauthorized
+      );
+    }
+
+    // Create assignmentData with 
+    const assignmentData = {
+      ...initialAssignmentData,
+      assignedAt: new Date(),
+    };
+
+    // add assignment to database
+    const addedAssignment = await addAssignmentToOrg(token, organizationId, assignmentData);
+
+    // Return a successful response.
+    return NextResponse.json(
+      { status: "success", message: "Assignment successfully.", data: addedAssignment},
+      { status: 201 } // Created
+    );
+
+  } catch (e) {
+    if (e instanceof FirebaseVerifyError || e instanceof FirestoreDatabaseError) {
+      return NextResponse.json(
+        { status: "fail", message: e.message },
+        { status: e.code }
+      );
+    }
+
+    // Genereic error handler
+    console.error("Error in create assignment route:", e);
+    return NextResponse.json(
+      { status: "error", message: "An internal server error occurred." },
+      { status: 500 } // Internal Server Error
+    );
+  }
+}
